conditional compilation for different os and arch
