+++ cmd/devtools/init.go
@@ -18,11 +18,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-type Binary struct {
-	Name string
-	Url  string
-}
-
 // initCmd represents the init command
 var initCmd = &cobra.Command{
 	Use:   "init",
@@ -55,15 +50,7 @@ func runInitialization() {
 	recreateMprocsFile(fullPath)
 	recreateJustfile(fullPath)
 
-	// TODO: make the binaries list configurable based on target os and arch
-	binaries := []Binary{
-		{"cometbft", "https://github.com/cometbft/cometbft/releases/download/v0.37.4/cometbft_0.37.4_darwin_arm64.tar.gz"},
-		{"astria-sequencer", "https://github.com/astriaorg/astria/releases/download/sequencer-v0.9.0/astria-sequencer-aarch64-apple-darwin.tar.gz"},
-		{"astria-composer", "https://github.com/astriaorg/astria/releases/download/composer-v0.4.0/astria-composer-aarch64-apple-darwin.tar.gz"},
-		{"astria-conductor", "https://github.com/astriaorg/astria/releases/download/conductor-v0.12.0/astria-conductor-aarch64-apple-darwin.tar.gz"},
-	}
-
-	for _, bin := range binaries {
+	for _, bin := range Binaries {
 
 		downloadAndUnpack(bin.Url, fullPath, bin.Name)
 	}
